requirements.txt
django==3.2.16
psycopg2-binary==2.9.5
djangorestframework==3.14.0
graphene-django==3.0.0
django_filter
django-graphql-jwt